pyproject.toml
[project]
name = "mas-template"
version = "0.1.0"
description = "Template project for building Multi-Agent Systems (MAS)"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "openai>=1.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.5.0",
    "openai-agents",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
    "langfuse",
    "nest_asyncio",
    "pydantic-ai[logfire]",
    "datasets",
    "nltk>=3.8.1",
    "fastapi>=0.112.0",
    "uvicorn>=0.29.0",
    "pydantic-settings>=2.7.1",
    "apscheduler>=3.11.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
